+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/export/response/AggregationWidgetExportResponseWriter.java
@@ -26,6 +26,7 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.ext.MessageBodyWriter;
 import jakarta.ws.rs.ext.Provider;
+import org.graylog.plugins.views.search.searchtypes.export.PivotResultExportResponse;
 import org.graylog2.rest.MoreMediaTypes;
 
 import java.io.IOException;
@@ -38,7 +39,7 @@ import java.util.List;
 
 @Provider
 @Produces({MoreMediaTypes.TEXT_CSV, MediaType.APPLICATION_JSON, MoreMediaTypes.APPLICATION_YAML})
-public class AggregationWidgetExportResponseWriter implements MessageBodyWriter<AggregationWidgetExportResponse> {
+public class AggregationWidgetExportResponseWriter implements MessageBodyWriter<PivotResultExportResponse> {
 
     private final ObjectMapper objectMapper;
     private final YAMLMapper yamlMapper;
@@ -53,11 +54,11 @@ public class AggregationWidgetExportResponseWriter implements MessageBodyWriter<
 
     @Override
     public boolean isWriteable(Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType) {
-        return aClass.equals(AggregationWidgetExportResponse.class);
+        return aClass.equals(PivotResultExportResponse.class);
     }
 
     @Override
-    public void writeTo(AggregationWidgetExportResponse widgetExportResponse, Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> multivaluedMap, OutputStream outputStream) throws IOException, WebApplicationException {
+    public void writeTo(PivotResultExportResponse widgetExportResponse, Class<?> aClass, Type type, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> multivaluedMap, OutputStream outputStream) throws IOException, WebApplicationException {
         switch (mediaType.toString()) {
             case MoreMediaTypes.TEXT_CSV -> writeCsv(widgetExportResponse, outputStream);
             case MediaType.APPLICATION_JSON -> objectMapper.writeValue(outputStream, widgetExportResponse);
@@ -66,13 +67,13 @@ public class AggregationWidgetExportResponseWriter implements MessageBodyWriter<
         }
     }
 
-    public static void writeCsv(final AggregationWidgetExportResponse widgetExportResponse,
+    public static void writeCsv(final PivotResultExportResponse widgetExportResponse,
                           final OutputStream outputStream) throws IOException {
         try (final CSVWriter csvWriter = new CSVWriter(new PrintWriter(outputStream, true, StandardCharsets.UTF_8))) {
 
             csvWriter.writeNext(widgetExportResponse.header().toArray(new String[0]));
-            for (List<String> row : widgetExportResponse.dataRows()) {
-                csvWriter.writeNext(row.toArray(new String[0]));
+            for (List<Object> row : widgetExportResponse.dataRows()) {
+                csvWriter.writeNext(row.stream().map(Object::toString).toList().toArray(new String[0]));
             }
 
         }
